requirements.txt
python-telegram-bot==13.15
flask
flask_sqlalchemy
requests
qrcode
pillow